requirements.txt
PyYAML
six
libkplug